requirements.txt
# Video Generation Pipeline Dependencies
# Install with: pip install -r requirements.txt

# Core PyTorch (install separately for your system)
# For CUDA 11.8: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# For CPU: pip install torch torchvision torchaudio

# Vision and Image Processing
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
pillow>=10.0.0
imageio>=2.33.0
imageio-ffmpeg>=0.4.9

# Hugging Face Transformers and Diffusers
transformers>=4.49.0
diffusers>=0.36.0
accelerate>=0.26.0
safetensors>=0.4.3

# Vision Models
timm>=0.9.0
numpy>=1.24.0

# Qwen VL (no longer needed - built into transformers)
# qwen-vl-utils>=0.0.8

# Scientific computing (for trajectory smoothing)
scipy>=1.10.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0

# IP-Adapter and SAM (for cross-frame consistency)
git+https://github.com/facebookresearch/segment-anything.git
git+https://github.com/tencent-ailab/IP-Adapter.git
# Set SAM_CHECKPOINT environment variable or place in models/
#
# Usage with IP-Adapter enabled:
# python scripts/generate_ip_keyframes.py --use-ip-adapter --ip-adapter-model models/ip-adapter_sd15.bin

# Development (optional)
# pytest>=7.4.0
# black>=23.9.0
# flake8>=6.0.0
